Returned functions to original scripts
