+++ scripts/util/general.py
